requirements.txt
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
fastapi==0.61.0
python-dotenv
pymongo
uvicorn
aiofiles
pymongo[srv]
python-jose[cryptography]
python-multipart